+++ src/whisker/src/insn32/op_imm_32.rs
@@ -17,7 +17,7 @@ pub fn parse_op_imm_32(_: &mut WhiskerHart, parcel: u32) -> Option<Instruction> 
 		),
 		SHIFT_LEFT_IMM_WORD => {
 			let shift_kind = extract_bits_32(itype.imm() as u32, 6, 11) as u8;
-			let shift_amt = extract_bits_32(itype.imm() as u32, 0, 5) as u32;
+			let shift_amt = extract_bits_32(itype.imm() as u32, 0, 5);
 
 			match shift_kind {
 				SHIFT_LOGICAL => Some(
@@ -33,7 +33,7 @@ pub fn parse_op_imm_32(_: &mut WhiskerHart, parcel: u32) -> Option<Instruction> 
 		}
 		SHIFT_RIGHT_IMM_WORD => {
 			let shift_kind = extract_bits_32(itype.imm() as u32, 6, 11) as u8;
-			let shift_amt = extract_bits_32(itype.imm() as u32, 0, 5) as u32;
+			let shift_amt = extract_bits_32(itype.imm() as u32, 0, 5);
 			match shift_kind {
 				SHIFT_LOGICAL => Some(
 					IntInstruction::ShiftRightLogicalImmediateWord {
